usb: dwc3: msm: Ensure USB2 SUSPHY bit is cleared during runstop 0
On SM6375/SM8150, it is observed that if the USB2PHYCFG SUSPHY bit is
set during clearing of RUN_STOP in device mode, then sometimes, the
DSTS Ctrl Dev halt bit it not being set soon and the timeout waiting
for it to be set is delaying run stop routing for more than 4 seconds
and during PIPO tests, this can result in causing next plugin
enumeration failure as the next plug-in's run stop setting doesn't go
through.

As per register dump observations, it is seen that UTMI clock is
suspended in fail case and as per SVE, it must be kept on during
this routine. So use the existing kretprobes to ensure UTMI clock
is kept on during run stop routine.

Ideally the wrapper takes care of setting and clearing this bit and
ensuring that it is cleared during runtime of core/gadget. So clear
this bit during run_stop(0) and leave it to msm_suspend for setting
it again.

Change-Id: I8878956c5dae82cc3566c8933407c5e10c0248dc
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>