msm: ep_pcie: Disable hot reset and ignore linkdown
Intel host tends to retrain the link if device doesn't send PM_Enter_L23
in time, leading to linkdown and hot reset in device side. The hot reset
here even results to PERST# timeout and device crash to PBL, which is
not expected.

Since we are in the entry of shutdown the link when linkdown happen, it
makes sense to ignore the linkdown. To avoid device crash because of
PERST# timeout, disable hot reset in pm_turnoff irq and reenable it when
we do linktrain again.

Change-Id: I87961faaae2a14c9a5e7d24543b416914076dd2e
Signed-off-by: Qiang Yu <quic_qianyu@quicinc.com>