msm: mhi_dev: Handle host wakeup in M3/D0
When a MHI WAKE request (that is request to bring MHI from M3 to M0) is
received while device is in D0, this request is being dropped as there
is no way to notify host about this event. This is leading to failure at
client drivers as they unable to wakeup mhi and perform write operation.

Fix the issue by waiting for D3hot in MHI before sending the wake
request:
    - If M0 is received while waiting, exit the function.
    - If D3hot/D3cold is received, send the wake request.
    - Else return failure.

Increase the timeout value of mhi_dev_write_channel() from 2s to 2.5s
to accommodate the waiting time for D state transition to D3hot/D3cold.

Change-Id: Idd58bb664d31bc1e5615119284c6cba924fe17b6
Signed-off-by: Sumit Kumar <quic_sumk@quicinc.com>