usb: dwc3: dwc3-msm-core: Handle notifications for multiple host controllers
When there are multiple USB controllers operating in host mode, there
is a possibility of receiving USB host notifications that belong to a
different controller. This can sometimes occur before the host is
initialized, leading to a NULL pointer dereference.

This patch adds necessary checks to ensure that only notifications
belonging to the matching controller are handled, preventing potential
NULL pointer dereferences and unwanted controller settings.

Change-Id: If557f98016314e183494071c68d839ae9b8d24cb
Signed-off-by: Faisal Hassan <quic_faisalh@quicinc.com>