pci: msm: Add i2c softdep for QPS615
In the pci-msm driver with the PCIe switch QPS615 attached, the
i2c_msm_geni driver needs to probe first to power up and set up
QPS615. Currently, there is no dependency on the i2c_msm_geni
driver in the pci-msm driver. As a result, the pci-msm driver
may probe before the i2c device, returning -EPROBE_DEFER. This
causes all drivers of devices attached to the PCIe switch to
also return -EPROBE_DEFER, or simply return an error if probe
defer is not supported.

The WLAN driver is not a platform driver, so it can't support
deferred probe. It requires the PCIe to be successfully probed
when CLD is loaded. Adding a dependency in DTS will cause a
deferred probe if the dependency handle is not ready. When the
pci-msm driver gets deferred, the WLAN will immediately fail.

To address this issue, a soft dependency on i2c_msm_geni is
added to pci_msm_drv. This ensures that the Linux kernel loads
and probes the i2c device before the PCIe device.

Change-Id: Ib587aec5d1a18e8c31a4d175d02bee06ce1c9e6a
Signed-off-by: Brandon Huang <quic_yanghuan@quicinc.com>
Signed-off-by: Jiani Liu <quic_jianil@quicinc.com>