usb: dwc3-msm: Add support to pass on HS port overcurrent info to host
The xHC has to set/clear OCA bit and set OCC bit of PORTSC
register on OverCurrent true and false conditions respectively.

On QCOM SoCs the xHC controllers are not capable of detecting
OCP on their ports. However there is a Qscratch register that can be
written to trigger the xHC to update the PORTSC bits.
Platforms may then be able to add external OC detection circuitry
and connect the detection output to a GPIO. This Qscratch register
can then be written up on receiving an interrupt on this GPIO when
it asserts.

Add support for the same based on USB fault GPIO probe and proper
QSCRATCH register update.

Change-Id: I042eb78a040b75dcd453ad16efd0beb131f9dc00
Signed-off-by: Sriram Dash <quic_sriramd@quicinc.com>