drivers: mhi: uci: Ensure channel unprepare on last fops release
MHI resources need to be freed only when the channel gets
removed. Currently, the MHI UCI driver increments ref_count
during probe and fops open calls. The `mhi_channel_prepare`
function is called only when ref_count is 2. However, once
all file descriptors are closed, the UCI driver does not
perform `mhi_channel_unprepare` because the ref_count is
non-zero during the release call. Currently, unprepare is
being called only when all ref_count reach zero.

This behavior causes the channel to be unprepared only when
the channel is destroyed, but not during the last file-
operation release call. This leads to an issue when the
client driver releases all file operation nodes and opens
them again. The MHI driver attempts to send a start command
for the channel, but since the stop command was never sent
during the previous fops release, the UCI driver fails to
open the fops.

To address this issue, the following changes are made:
- Modify the release call to check if it is the last file
  descriptor being closed and perform the unprepare operation if so.
- Ensure that start and stop commands for the channel are correctly paired.

Change-Id: Id1da13a5bd9754448707e84e83a12237e135c58b
Signed-off-by: Vivek Pernamitta <quic_vpernami@quicinc.com>